refactor: adjust the file location
