+++ src/articles/comments/responses/delete-comment.response.ts
@@ -1,0 +1,17 @@
+import { ApiProperty } from "@nestjs/swagger";
+
+export class DeleteCommentResponse {
+  @ApiProperty({
+    type: "number",
+    description: "HTTP 回應代碼",
+    example: 200,
+  })
+  public readonly statusCode: number;
+
+  @ApiProperty({
+    type: "string",
+    description: "刪除成功回應",
+    example: "刪除成功",
+  })
+  public readonly message: string;
+}
