+++ src/articles/comments/responses/select-like-comment.response.ts
